requirements.txt
python-dotenv
requests
urllib3<2
flask
flask-sqlalchemy
